working changes : properly saving database for time, distance and workoutNumber for All_TIme showing correct data on Profile Activity TO DO : (need to implement average weekly data and show live data on UI) (need to work on saving points in service rather than on Activity)

+++ app/src/main/java/edu/sjsu/anis/alphafitness/WorkoutFragment.java
@@ -270,7 +270,7 @@ public class WorkoutFragment extends Fragment {
             @Override
             public void handleMessage(Message msg) {
 
-                double calDistance = (double) msg.obj;
+                float calDistance = (float) msg.obj;
 
 //                Toast.makeText(getActivity(), "Distance: " + distance, Toast.LENGTH_SHORT).show();
                 distanceUI.setText(String.format("%.3f", calDistance));
